+++ src/Query/Testing/RegionBuilderFake.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Katalam\OnOfficeAdapter\Query\Testing;
+
+class RegionBuilderFake extends BaseFake
+{
+}