+++ src/Repositories/SettingRepository.php
@@ -2,10 +2,11 @@
 
 namespace Katalam\OnOfficeAdapter\Repositories;
 
+use Katalam\OnOfficeAdapter\Query\RegionBuilder;
 use Katalam\OnOfficeAdapter\Query\UserBuilder;
 use Katalam\OnOfficeAdapter\Services\OnOfficeService;
 
-class UserRepository
+class SettingRepository
 {
     public function __construct(
         private readonly OnOfficeService $onOfficeService,
@@ -15,8 +16,16 @@ class UserRepository
     /**
      * Returns a new user builder instance.
      */
-    public function query(): UserBuilder
+    public function users(): UserBuilder
     {
         return new UserBuilder($this->onOfficeService);
+    }
+
+    /**
+     * Returns a new regions builder instance.
+     */
+    public function regions(): RegionBuilder
+    {
+        return new RegionBuilder($this->onOfficeService);
     }
 }